Improve conversion utilities
