--- packages/rrule-generator/src/translations/index.js
@@ -1,4 +1,0 @@
-import english from './english';
-import german from './german';
-
-export default { english, german };
