--- packages/rrule-generator/src/utils/computeRRule/fromString/computeDailyInterval.js
@@ -1,9 +1,0 @@
-const computeDailyInterval = (data, rruleObj) => {
-  if (rruleObj.freq !== 3) {
-    return data.repeat.daily.interval;
-  }
-
-  return rruleObj.interval;
-};
-
-export default computeDailyInterval;
