+++ packages/rrule-generator/src/utils/computeRRule/fromString/computeDailyInterval.ts
@@ -1,0 +1,11 @@
+import { ComputeRule } from './types';
+
+const computeDailyInterval: ComputeRule<number> = (data, rruleObj) => {
+	if (rruleObj.freq !== 3) {
+		return data?.repeat?.daily?.interval;
+	}
+
+	return rruleObj.interval;
+};
+
+export default computeDailyInterval;
