--- packages/rrule-generator/src/utils/computeRRule/fromString/computeEndAfter.js
@@ -1,9 +1,0 @@
-const computeEndAfter = (data, rruleObj) => {
-  if (!rruleObj.count && rruleObj.count !== 0) {
-    return data.end.after;
-  }
-
-  return rruleObj.count;
-};
-
-export default computeEndAfter;
