+++ packages/rrule-generator/src/utils/computeRRule/fromString/computeEndDate.ts
@@ -1,0 +1,11 @@
+import { ComputeRule } from './types';
+
+const computeEndOnDate: ComputeRule<Date> = (data, rruleObj) => {
+	if (!rruleObj.until) {
+		return data?.end?.date;
+	}
+
+	return rruleObj.until;
+};
+
+export default computeEndOnDate;
