--- packages/rrule-generator/src/utils/computeRRule/fromString/computeEndOnDate.js
@@ -1,9 +1,0 @@
-const computeEndOnDate = (data, rruleObj) => {
-  if (!rruleObj.until) {
-    return data.end.date;
-  }
-
-  return rruleObj.until;
-};
-
-export default computeEndOnDate;
