--- packages/rrule-generator/src/utils/computeRRule/fromString/computeHourlyInterval.js
@@ -1,9 +1,0 @@
-const computeHourlyInterval = (data, rruleObj) => {
-  if (rruleObj.freq !== 4) {
-    return data.repeat.daily.interval;
-  }
-
-  return rruleObj.interval;
-};
-
-export default computeHourlyInterval;
