+++ packages/rrule-generator/src/utils/computeRRule/fromString/computeHourlyInterval.ts
@@ -1,0 +1,11 @@
+import { ComputeRule } from './types';
+
+const computeHourlyInterval: ComputeRule<number> = (data, rruleObj) => {
+	if (rruleObj.freq !== 4) {
+		return data?.repeat?.daily?.interval;
+	}
+
+	return rruleObj.interval;
+};
+
+export default computeHourlyInterval;
