--- packages/rrule-generator/src/utils/computeRRule/fromString/computeMonthlyInterval.js
@@ -1,9 +1,0 @@
-const computeMonthlyInterval = (data, rruleObj) => {
-  if (rruleObj.freq !== 1) {
-    return data.repeat.monthly.interval;
-  }
-
-  return rruleObj.interval;
-};
-
-export default computeMonthlyInterval;
