+++ packages/rrule-generator/src/utils/computeRRule/fromString/computeMonthlyInterval.ts
@@ -1,0 +1,11 @@
+import { ComputeRule } from './types';
+
+const computeMonthlyInterval: ComputeRule<number> = (data, rruleObj) => {
+	if (rruleObj.freq !== 1) {
+		return data?.repeat?.monthly?.interval;
+	}
+
+	return rruleObj.interval;
+};
+
+export default computeMonthlyInterval;
