--- packages/rrule-generator/src/utils/computeRRule/fromString/computeStartOnDate.js
@@ -1,8 +1,0 @@
-const computeStartOnDate = (data, rruleObj) => {
-  if (!rruleObj.dtstart) {
-    return data?.start?.date;
-  }
-
-  return rruleObj.dtstart;
-};
-export default computeStartOnDate;
