--- packages/rrule-generator/src/utils/computeRRule/fromString/computeWeeklyInterval.js
@@ -1,9 +1,0 @@
-const computeWeeklyInterval = (data, rruleObj) => {
-  if (rruleObj.freq !== 2) {
-    return data.repeat.weekly.interval;
-  }
-
-  return rruleObj.interval;
-};
-
-export default computeWeeklyInterval;
