+++ packages/rrule-generator/src/utils/computeRRule/fromString/computeWeeklyInterval.ts
@@ -1,0 +1,11 @@
+import { ComputeRule } from './types';
+
+const computeWeeklyInterval: ComputeRule<number> = (data, rruleObj) => {
+	if (rruleObj.freq !== 2) {
+		return data?.repeat?.weekly?.interval;
+	}
+
+	return rruleObj.interval;
+};
+
+export default computeWeeklyInterval;
