--- packages/rrule-generator/src/utils/computeRRule/fromString/computeYearlyMode.js
@@ -1,13 +1,0 @@
-const computeYearlyMode = (data, rruleObj) => {
-  if (rruleObj.freq !== 0 || !rruleObj.bymonth) {
-    return data.repeat.yearly.mode;
-  }
-
-  if (rruleObj.bymonthday) {
-    return 'on';
-  }
-
-  return 'on the';
-};
-
-export default computeYearlyMode;
