+++ packages/rrule-generator/src/utils/computeRRule/fromString/types.ts
@@ -1,0 +1,4 @@
+import { Options } from 'rrule';
+import { RRuleState } from '../../../state';
+
+export type ComputeRule<T = any> = (data: RRuleState, rRuleObj: Partial<Options>) => T;
