+++ packages/rrule-generator/src/utils/computeRRule/toString/computeYearlyOn.ts
@@ -1,10 +1,10 @@
 import { Options } from 'rrule';
 
-import { MONTHS } from '../../../constants/index';
+import { MONTHS } from '../../../constants';
 import { RRuleState } from '../../../state';
 
 const computeYearlyOn = (on: RRuleState['repeat']['yearly']['on']): Partial<Options> => ({
-	bymonth: MONTHS.indexOf(on.month) + 1,
+	bymonth: Object.keys(MONTHS).indexOf(on.month) + 1,
 	bymonthday: on.day as number,
 });
